added FitTotECoeffs_Heteronuclear.jl and CompareHeteronuclearFits.jl
Preliminary scrips for fitting and validating the total energy empirical coefficients for heteronuclear diatomic cases have been added.

The current version of the polarization Lagrangian seems to work very well for homonuclear cases but not for heteronuclear cases, since there are plenty of discontinuities.

Currently, I am working on finding and testing different small reformulations of the polarization Lagrangian that fix this issue.