Code readability improvements
Overall improvements in how the code is structured, using a more object-oriented approach, instead of using the matrices "atoms_data" and "cloud_data".

The electron densities are now fitted using the same level of theory as the reference calculations. Six s-type Gaussian functions are used to fit each of the electron shells of the atoms in the first three rows of the periodic table.